Made delaunay triangulations almost work.
There's a unit test we're not passing now for pyVor.predicates.incircle.
In the best case, as soon as we get that to pass, the tests for
DelaunayTriangulation will start to pass immediately as well.

I am committing in this unfinished state because I have to do
Thanksgiving things, and so I am about to be much less available
to work on this. But maybe one of you really feels like fixing
the incircle predicate on thanksgiving.